Change int to long

+++ DJ_App/src/main/java/com/google/sps/data/Event.java
@@ -4,14 +4,14 @@ package com.google.sps.data;
 public final class Event {
 
   private final long tokenID;
-  private final int eventID;
+  private final long eventID;
   private final String djName;
   private final String eventName;
   private final String location;
   private final String timeRange;
   private final String eventDescription;
 
-  public Event (long ID, int eventID, String djName, String eventName, String location, String timeRange, String eventDescription) {
+  public Event (long ID, long eventID, String djName, String eventName, String location, String timeRange, String eventDescription) {
     this.tokenID = ID;
     this.eventID = eventID;
     this.djName = djName;
